requirements.txt
Flask
Flask-Cors
SQLAlchemy
PyYAML
Flask-SQLAlchemy